requirements.txt
pefile~=2024.8.26
requests~=2.32.4
python-dotenv~=1.1.1